fix(sidebar): bound shared git probes without double-counting spawns
Use the bounded process runner for both refresh lanes. Keep raw output for
multiline and NUL consumers and one spawn count per command attempt.
